Add scanletter alias to .bashrc.d/10aliases.
This makes it easy for me to scan letter-size pages:

  $ scanletter > scan.tiff

It's based on `scanimage` (media-gfx/sane-backends) on Gentoo, which
uses the epson2 sane backend and the iscan gt f720 plugin for my Epson
V300.  To emerge on Gentoo, add

  SANE_BACKENDS="epson2"

to your `/etc/make.conf`, and emerge `media-gfx/xsane` and
`media-gfx/iscan-plugin-gt-f720`.  You may not need all of xsane for
this alias, but it's nice to have for scanning strangely shaped
objects.